reorganized StreamRefs into implementation folder
